+++ tests/views.py
@@ -2,12 +2,15 @@ from rest_framework_json_api.views import ModelViewSet
 from tests.models import (
     BasicModel,
     ForeignKeySource,
+    ForeignKeyTarget,
     ManyToManySource,
     NestedRelatedSource,
 )
 from tests.serializers import (
     BasicModelSerializer,
     ForeignKeySourceSerializer,
+    ForeignKeySourcetHyperlinkedSerializer,
+    ForeignKeyTargetSerializer,
     ManyToManySourceSerializer,
     NestedRelatedSourceSerializer,
 )
@@ -22,6 +25,18 @@ class BasicModelViewSet(ModelViewSet):
 class ForeignKeySourceViewSet(ModelViewSet):
     serializer_class = ForeignKeySourceSerializer
     queryset = ForeignKeySource.objects.all()
+    ordering = ["name"]
+
+
+class ForeignKeySourcetHyperlinkedViewSet(ModelViewSet):
+    serializer_class = ForeignKeySourcetHyperlinkedSerializer
+    queryset = ForeignKeySource.objects.all()
+    ordering = ["name"]
+
+
+class ForeignKeyTargetViewSet(ModelViewSet):
+    serializer_class = ForeignKeyTargetSerializer
+    queryset = ForeignKeyTarget.objects.all()
     ordering = ["name"]